+++ Gruntfile.js
@@ -6,6 +6,7 @@ src.util = [
   'js/widget.parser.js',
   'js/widget.util.js',
   'js/widget.format.js',
+  'js/widget.validate.js',
   'js/parser/argument.js',
   'js/parser/inlineFunction.js',
   'js/parser/localize.js',