+++ js/global.js
@@ -28,8 +28,8 @@ if (!String.prototype.endsWith) {
 
 // http://stackoverflow.com/questions/46155/validate-email-address-in-javascript
 function validateEmail( email ) {
-  var re = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-  return re.test(email);
+  console.error( 'deprecation warning - use widet.validate.email() instead' );
+  return widget.validate.email( email );
 }