+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 --
 * tabChange event on tabGroup widget, tabWillChange events on the tab manager
 * URL argument template using format ?{key}
+* Deprecated global validateEmail() in favor of widget.validate.email()
 
 
 0.7.2 (2016.03.10)